+++ src/UltraSinger.py
@@ -8,7 +8,9 @@ import re
 
 import Levenshtein
 import librosa
+
 from tqdm import tqdm
+import soundfile as sf
 
 from modules import os_helper
 from modules.Audio.denoise import ffmpeg_reduce_noise
@@ -17,7 +19,7 @@ from modules.Audio.vocal_chunks import (
     export_chunks_from_transcribed_data,
     export_chunks_from_ultrastar_data,
 )
-from modules.Audio.silence_processing import remove_silence_from_transcription_data
+from modules.Audio.silence_processing import remove_silence_from_transcription_data, get_silence_sections
 from modules.csv_handler import export_transcribed_data_to_csv
 from modules.Audio.convert_audio import convert_audio_to_mono_wav, convert_wav_to_mp3
 from modules.Audio.youtube import (
@@ -131,6 +133,10 @@ def add_hyphen_to_data(transcribed_data: list[TranscribedData], hyphen_words: li
                 dup.end = next_start
                 dup.word = hyphen_words[i][hyphenated_word_index]
                 dup.is_hyphen = True
+                if hyphenated_word_index == len(hyphen_words[i]) - 1:
+                    dup.is_word_end = True
+                else:
+                    dup.is_word_end = False
                 new_data.append(dup)
 
     return new_data
@@ -327,7 +333,7 @@ def run() -> None:
         ) = infos_from_audio_input_file()
 
     cache_path = os.path.join(song_output, "cache")
-    settings.mono_audio_path = os.path.join(
+    settings.processing_audio_path = os.path.join(
         cache_path, basename_without_ext + ".wav"
     )
     os_helper.create_folder(cache_path)
@@ -353,7 +359,15 @@ def run() -> None:
         cache_path, basename_without_ext + "_mono.wav"
     )
     convert_audio_to_mono_wav(denoised_output_path, mono_output_path)
-    settings.mono_audio_path = mono_output_path
+
+    # Mute silence sections
+    mute_output_path = os.path.join(
+        cache_path, basename_without_ext + "_mute.wav"
+    )
+    mute_no_singing_parts(mono_output_path, mute_output_path)
+
+    # Define the audio file to process
+    settings.processing_audio_path = mute_output_path
 
     # Audio transcription
     transcribed_data = None
@@ -364,14 +378,15 @@ def run() -> None:
             language = detected_language
 
         remove_unecessary_punctuations(transcribed_data)
-        transcribed_data = remove_silence_from_transcription_data(
-            settings.mono_audio_path, transcribed_data
-        )
 
         if settings.hyphenation:
             hyphen_words = hyphenate_each_word(language, transcribed_data)
             if hyphen_words is not None:
                 transcribed_data = add_hyphen_to_data(transcribed_data, hyphen_words)
+
+        transcribed_data = remove_silence_from_transcription_data(
+            settings.processing_audio_path, transcribed_data
+        )
 
         # todo: do we need to correct words?
         # lyric = 'input/faber_lyric.txt'
@@ -395,7 +410,8 @@ def run() -> None:
     # Create plot
     if settings.create_plot:
         vocals_path = os.path.join(audio_separation_path, "vocals.wav")
-        plot_spectrogram(vocals_path, song_output)
+        plot_spectrogram(vocals_path, song_output, "vocals.wav")
+        plot_spectrogram(settings.processing_audio_path, song_output, "processing audio")
         plot(pitched_data, song_output, transcribed_data, midi_notes)
 
     # Write Ultrastar txt
@@ -436,6 +452,27 @@ def run() -> None:
     print_support()
 
 
+def mute_no_singing_parts(mono_output_path, mute_output_path):
+    print(
+        f"{ULTRASINGER_HEAD} Mute audio parts with no singing"
+    )
+    silence_sections = get_silence_sections(mono_output_path)
+    y, sr = librosa.load(mono_output_path, sr=None)
+    # Mute the parts of the audio with no singing
+    for i in silence_sections:
+        # Define the time range to mute
+
+        start_time = i[0]  # Start time in seconds
+        end_time = i[1]  # End time in seconds
+
+        # Convert time to sample indices
+        start_sample = int(start_time * sr)
+        end_sample = int(end_time * sr)
+
+        y[start_sample:end_sample] = 0
+    sf.write(mute_output_path, y, sr)
+
+
 def get_unused_song_output_dir(path: str) -> str:
     """Get an unused song output dir"""
     # check if dir exists and add (i) if it does
@@ -461,7 +498,7 @@ def transcribe_audio() -> (str, list[TranscribedData]):
     if settings.transcriber == "whisper":
         device = "cpu" if settings.force_whisper_cpu else settings.pytorch_device
         transcribed_data, detected_language = transcribe_with_whisper(
-            settings.mono_audio_path,
+            settings.processing_audio_path,
             settings.whisper_model,
             device,
             settings.whisper_align_model,
@@ -773,7 +810,7 @@ def pitch_audio(is_audio: bool, transcribed_data: list[TranscribedData], ultrast
     # midi_notes = pitch_each_chunk_with_crepe(chunk_folder_name)
     device = "cpu" if settings.force_crepe_cpu else settings.tensorflow_device
     pitched_data = get_pitch_with_crepe_file(
-        settings.mono_audio_path,
+        settings.processing_audio_path,
         settings.crepe_model_capacity,
         settings.crepe_step_size,
         device,
@@ -811,14 +848,13 @@ def create_audio_chunks(
     if is_audio:  # and csv
         csv_filename = os.path.join(audio_chunks_path, "_chunks.csv")
         export_chunks_from_transcribed_data(
-            settings.mono_audio_path, transcribed_data, audio_chunks_path
+            settings.processing_audio_path, transcribed_data, audio_chunks_path
         )
         export_transcribed_data_to_csv(transcribed_data, csv_filename)
     else:
         export_chunks_from_ultrastar_data(
             ultrastar_audio_input_path, ultrastar_class, audio_chunks_path
         )
-
 
 def denoise_vocal_audio(input_path: str, output_path: str) -> None:
     """Denoise vocal audio"""